I've refactored the Rust codebase to improve modularity and organization.
This commit refactors the Rust codebase to follow the recommended project structure. This includes:

- Separating the codebase into logical modules (agent, plugins, storage, metrics, etc.).
- Creating `mod.rs` files for each module to declare its contents.
- Moving core components to their appropriate locations.
- Updating `Cargo.toml` with new features and dependencies.
- Fixing all compilation errors and warnings.
