+++ rust/Cargo.toml
@@ -1,0 +1,81 @@
+[package]
+name = "reasoning-agent"
+version = "0.2.0"
+edition = "2021"
+rust-version = "1.70"
+authors = ["Your Name <email@example.com>"]
+license = "MIT"
+description = "A sophisticated reasoning agent with knowledge graphs and LLM integration"
+repository = "https://github.com/Loofy147/Rust"
+keywords = ["ai", "llm", "knowledge-graph", "reasoning", "agent"]
+categories = ["science", "algorithms", "web-programming"]
+
+[lib]
+name = "reasoning_agent"
+path = "src/lib.rs"
+
+[[bin]]
+name = "reasoning-agent"
+path = "src/main.rs"
+required-features = ["cli"]
+
+[dependencies]
+# Core async runtime
+tokio = { version = "1.35", features = ["full"] }
+async-trait = "0.1"
+
+# Serialization
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+# Error handling
+anyhow = "1.0"
+thiserror = "1.0"
+
+# Logging and tracing
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+
+# Configuration
+config = { version = "0.13", optional = true }
+
+# CLI
+clap = { version = "4.0", features = ["derive"], optional = true }
+
+# HTTP server (optional)
+axum = { version = "0.7", optional = true }
+tower = { version = "0.4", optional = true }
+tower-http = { version = "0.5", features = ["cors", "trace"], optional = true }
+
+# Database (optional)
+sqlx = { version = "0.7", features = ["runtime-tokio-rustls", "sqlite", "chrono"], optional = true }
+
+# External API clients (optional)
+reqwest = { version = "0.11", features = ["json"], optional = true }
+
+# Concurrent data structures
+dashmap = "5.5"
+
+# Metrics
+metrics = { version = "0.21", optional = true }
+uuid = { version = "1.8", features = ["v4"] }
+
+[features]
+default = ["cli"]
+cli = ["clap", "config"]
+api = ["axum", "tower", "tower-http"]
+database = ["sqlx", "sqlite"]
+llm-apis = ["reqwest"]
+metrics = ["dep:metrics", "prometheus"]
+full = ["cli", "api", "database", "llm-apis", "metrics"]
+sqlite = []
+prometheus = []
+
+[dev-dependencies]
+tokio-test = "0.4"
+tempfile = "3.0"
+wiremock = "0.5"
+
+[profile.release]
+lto = true
+codegen-units = 1
