+++ rust/src/agent/core.rs
@@ -1,0 +1,29 @@
+use serde::{Deserialize, Serialize};
+use crate::storage::memory::KnowledgeGraph;
+use crate::storage::vector_store::VectorStore;
+use crate::plugins::traits::{LLMPlugin, PromptBuilderPlugin};
+use crate::metrics::collector::MetricsCollector;
+use anyhow::Result;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ReasoningAgent {
+    pub id: String,
+}
+
+impl ReasoningAgent {
+    pub fn new(
+        _knowledge_graph: impl KnowledgeGraph + 'static,
+        _vector_store: impl VectorStore + 'static,
+        _llm_plugin: Box<dyn LLMPlugin>,
+        _prompt_builder: Box<dyn PromptBuilderPlugin>,
+        _metrics: MetricsCollector,
+    ) -> Result<Self> {
+        Ok(Self {
+            id: "agent-1".to_string(),
+        })
+    }
+
+    pub async fn process_task(&self, task: &str) -> Result<String> {
+        Ok(format!("Processed task: {}", task))
+    }
+}
