+++ rust/src/agent/mod.rs
@@ -1,0 +1,8 @@
+//! Core reasoning agent implementation
+
+pub mod core;
+pub mod config;
+pub mod prompt;
+
+pub use core::ReasoningAgent;
+pub use config::AgentConfig;
