+++ rust/src/api/handlers.rs
@@ -1,0 +1,2 @@
+// This file is intentionally left blank for now.
+// It will be populated with the API handlers.
