+++ rust/src/api/mod.rs
@@ -1,0 +1,2 @@
+pub mod handlers;
+pub mod models;
