+++ rust/src/metrics/mod.rs
@@ -1,0 +1,9 @@
+//! Metrics collection and export
+
+pub mod collector;
+pub mod exporters;
+
+pub use collector::MetricsCollector;
+#[cfg(feature = "prometheus")]
+pub use exporters::PrometheusExporter;
+pub use exporters::JsonExporter;
