+++ rust/src/plugins/mod.rs
@@ -1,0 +1,7 @@
+//! Plugin system for extensible functionality
+
+pub mod traits;
+pub mod llm;
+pub mod prompt;
+
+pub use traits::{LLMPlugin, PromptBuilderPlugin};
