+++ rust/src/plugins/prompt/mod.rs
@@ -1,0 +1,2 @@
+pub mod default;
+pub mod enhanced;
