+++ rust/src/utils/mod.rs
@@ -1,0 +1,5 @@
+//! Utility functions and common types
+
+pub mod error;
+
+pub use error::{Result, AgentError};
