--- src/types.rs
@@ -1,46 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct KnowledgeEntity {
-    pub id: String,
-    pub entity_type: String,
-    pub properties: Vec<(String, String)>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct VectorSimilarity {
-    pub id: String,
-    pub content: String,
-    pub similarity_score: f32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PromptContext {
-    pub task: String,
-    pub kg_entities: Vec<KnowledgeEntity>,
-    pub similar_vectors: Vec<VectorSimilarity>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LLMRequest {
-    pub prompt: String,
-    pub max_tokens: Option<u32>,
-    pub temperature: Option<f32>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LLMResponse {
-    pub content: String,
-    pub tokens_used: Option<u32>,
-    pub model: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TaskResult {
-    pub task_id: String,
-    pub task: String,
-    pub result: String,
-    pub duration_ms: u64,
-    pub kg_entities_used: usize,
-    pub similar_vectors_found: usize,
-}
